requirements.txt
# ---------------------------
# File: requirements.txt
# ---------------------------
faiss-gpu
mypy>=1.11.0
numpy>=1.26.4
python-dotenv>=1.0.1
pytest>=8.3.2
sentence-transformers>=3.0.1
setuptools>=70.0.0
striprtf>=0.0.26
tqdm>=4.66.5
wheel>=0.43.0
pydantic>=2.8.2
pyperclip>=1.9.0
pdfplumber>=0.11.4
python-docx>=1.1.2